feat(miner-foundation): extract scoring preview/model into gittensory-engine
Ports the deterministic parts of src/scoring/{preview,model,pending-pr-scenarios}.ts
into packages/gittensory-engine/src/scoring/, giving the miner the same
score-preview machinery the gittensory_preview_local_pr_score MCP tool already
calls server-side, so a miner can preview its own PR score fully locally (#2282).

The D1/GitHub-fetching halves of model.ts and pending-pr-scenarios.ts stay in
src/, which now imports its pure constants/classifiers back from the engine;
those functions have no use to a local miner (no D1/App-token access) and
depend on Cloudflare-specific bindings the engine package's isolated tsconfig
cannot resolve. src/scoring/preview.ts is now a full re-export shim since the
whole file was already side-effect-free. Every existing scoring test suite
passes unmodified against the new shims.
